updated repo

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,12 +8,11 @@
 
 size_t print_list(const list_t *h)
 {
-	int num_of_node = 0;
-	/* h = malloc(sizeof(const list_t *)); */
+	size_t num_of_node = 0;
 
 	while (h != NULL)
 	{
-		(h->str != NULL) ? printf("[%u]%s\n", h->len,  h->str) : printf("[0] (nill)\n");
+		(h->str != NULL) ? printf("[%d]%s\n", h->len,  h->str) : printf("[0] (nill)\n");
 		h = h->next;
 		num_of_node++;
 	}
